Moved TableCommitMetaStoreWorker to nessie-services
