+++ servers/services/src/main/java/com/dremio/nessie/services/rest/TreeResource.java
@@ -69,7 +69,7 @@ import com.google.common.collect.ImmutableList;
 public class TreeResource extends BaseResource implements TreeApi {
 
   @Inject
-  protected TreeResource(ServerConfig config, Principal principal,
+  public TreeResource(ServerConfig config, Principal principal,
       VersionStore<Contents, CommitMeta> store) {
     super(config, principal, store);
   }
